requirements.txt
fire
pillow
python-telegram-bot
